add service 5+6+7 to the task

+++ Flight Management System03/Models/Flight.cs
@@ -21,6 +21,6 @@ namespace Flight_Management_System03.Models
 
         public string status { get; set; } //default value, Flight status: Scheduled | Departed | Cancelled
 
-        public double flightDuration { get; set; } //user input
+        public int flightDuration { get; set; } //user input
     }
 }
